move user profile tests around, test a profile with tags
